Add files via upload
What is a webhook ? A webhook is a way for one application to send real-time data to another application when a specific event occurs. It is a server-to-server notification mechanism for sending real time Data. What is an Event ? Understand from this scenario, lets say the payment is successfully done and the payment gateway captures the payment. Here the event is payment capture. Based on this event we will call the Callback API and send the data.